requirements.txt
streamlit
tensorflow==2.8.0
protobuf==3.20.3
gdown
numpy
Pillow